EE-6713: JMS number of consumers should be adapted in case it is greater than available cores.
