+++ src/test/resources/multiconsumer/jms-default-multi-consumers.xml
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<mule xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      xmlns:jms="http://www.mulesoft.org/schema/mule/jms"
+      xmlns:test="http://www.mulesoft.org/schema/mule/test"
+      xmlns="http://www.mulesoft.org/schema/mule/core"
+      xsi:schemaLocation="http://www.mulesoft.org/schema/mule/core http://www.mulesoft.org/schema/mule/core/current/mule.xsd
+             http://www.mulesoft.org/schema/mule/test http://www.mulesoft.org/schema/mule/test/current/mule-test.xsd
+      http://www.mulesoft.org/schema/mule/jms http://www.mulesoft.org/schema/mule/jms/current/mule-jms.xsd">
+
+    <flow name="publisher">
+        <set-payload value="#[payload]" mimeType="application/json"/>
+        <jms:publish config-ref="config" destination="#[vars.destination]" destinationType="#[vars.destinationType]"/>
+        <logger level="DEBUG" message="#[payload]"/>
+    </flow>
+
+    <flow name="listener">
+        <jms:listener config-ref="config" destination="${destination}" ackMode="MANUAL" />
+        <logger level="DEBUG" message="#[payload]"/>
+        <flow-ref name="saveMessage" />
+    </flow>
+
+    <flow name="topicListener" initialState="stopped">
+        <jms:listener config-ref="config" destination="${topicDestination}" ackMode="MANUAL">
+            <jms:consumer-type>
+                <jms:topic-consumer/>
+            </jms:consumer-type>
+        </jms:listener>
+        <logger level="DEBUG" message="#[payload]"/>
+        <flow-ref name="saveMessage" />
+    </flow>
+
+    <flow name="saveMessage">
+        <custom-processor class="org.mule.extensions.jms.test.JmsMessageStorage"/>
+    </flow>
+
+    <flow name="doManualAck">
+        <logger level="DEBUG" message="ACKNOWLEDGING SESSION"/>
+        <jms:ack ackId="#[payload]"/>
+    </flow>
+
+    <flow name="recoverSession">
+        <logger level="DEBUG" message="RECOVERING SESSION"/>
+        <jms:recover-session ackId="#[payload]"/>
+    </flow>
+</mule>
